mutators/generators: fix wraparound at numPixels()
The moving and offset mutators would clamp offsets modulo numPixels().
This is suboptimal as it limits the possibilities for valid offsets.
Modulo wrapping is only required to make sure no invalid pixel offsets
are used. This should be ensured by the generators instead. Right now
the only generator requiring the wrapping is the 'CurrentColor' one.

Signed-off-by: Stefan Nuernberger <kabelfrickler@gmail.com>